# public/openapi.yaml
openapi: 3.0.0
info:
  title: Employee Attendance API
  version: 1.0.0
  description: REST API for managing employees, attendance, and daily attendance reports
servers:
  - url: http://localhost:8000
    description: Local development (Sail on port 8000)
paths:
  /api/v1/attendance/check-in:
    post:
      tags:
        - Attendance
      summary: 'Check in an employee'
      operationId: 687a4f2a1df43ca1081c615e110403db
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employee_id:
                  type: integer
                  nullable: true
                  example: 1
                employee_identifier:
                  type: string
                  nullable: true
                  example: EMP-001
                notes:
                  type: string
                  nullable: true
                  example: Arrived on time
      responses:
        '201':
          description: 'Checked in'
        '401':
          description: Unauthenticated
        '409':
          description: Conflict
        '422':
          description: 'Validation error'
  /api/v1/attendance/check-out:
    post:
      tags:
        - Attendance
      summary: 'Check out an employee'
      operationId: e105a11f0b3d48a42178e29d33328682
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employee_id:
                  type: integer
                  nullable: true
                  example: 1
                employee_identifier:
                  type: string
                  nullable: true
                  example: EMP-001
                notes:
                  type: string
                  nullable: true
                  example: Left early for appointment
      responses:
        '200':
          description: 'Checked out'
        '401':
          description: Unauthenticated
        '409':
          description: Conflict
        '422':
          description: 'Validation error'
  /api/v1/auth/register:
    post:
      tags:
        - Auth
      summary: 'Register a new user'
      operationId: 7ca6fc7989202054d29707fef91a2d4a
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - email
                - password
                - password_confirmation
              properties:
                name:
                  type: string
                  example: 'John Doe'
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password
                password_confirmation:
                  type: string
                  format: password
                  example: password
              type: object
      responses:
        '201':
          description: 'User registered successfully'
        '422':
          description: 'Validation error'
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: Login
      operationId: 9d0bb5c770cbec19b0b9566606ef4adb
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password
              type: object
      responses:
        '200':
          description: 'Login successful'
        '401':
          description: 'Invalid credentials'
  /api/v1/auth/logout:
    post:
      tags:
        - Auth
      summary: Logout
      operationId: 3a27a9e02ceb521eb3f98c8d2394aef9
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 'Logout successful'
        '401':
          description: Unauthenticated
  /api/v1/auth/change-password:
    post:
      tags:
        - Auth
      summary: 'Change password (authenticated user)'
      operationId: 6f96a3b0b2f84701a445ef61a34e0cbe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - current_password
                - password
                - password_confirmation
              properties:
                current_password:
                  type: string
                  format: password
                  example: old-password
                password:
                  type: string
                  format: password
                  example: new-password
                password_confirmation:
                  type: string
                  format: password
                  example: new-password
      responses:
        '200':
          description: 'Password changed successfully'
        '401':
          description: 'Unauthenticated'
        '422':
          description: 'Validation error / wrong current password'
  /api/v1/auth/forgot-password:
    post:
      tags:
        - Auth
      summary: 'Send password reset link'
      operationId: 60d704a0913808970fd9d7baadde0662
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          description: 'Reset link sent (if email exists)'
        '422':
          description: 'Validation error'
  /api/v1/auth/reset-password:
    post:
      tags:
        - Auth
      summary: 'Reset password'
      operationId: b10cb4a66c8bb98c8db204c20e546148
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - email
                - password
                - password_confirmation
              properties:
                token:
                  type: string
                  example: reset-token-from-email
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: new-password
                password_confirmation:
                  type: string
                  format: password
                  example: new-password
      responses:
        '200':
          description: 'Password reset successful'
        '422':
          description: 'Invalid token / validation error'
  /api/v1/employees:
    get:
      tags:
        - Employees
      summary: 'List employees'
      operationId: 326f2e86fa40a9ba88b1cb5f2d4b62b8
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 'Employees list'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
    post:
      tags:
        - Employees
      summary: 'Create employee'
      operationId: 4a10b3ada27736863f2586c56b9c19ec
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - employee_identifier
              properties:
                name:
                  type: string
                  example: Jane Doe
                email:
                  type: string
                  format: email
                  example: employee@example.com
                employee_identifier:
                  type: string
                  example: EMP-001
                phone_number:
                  type: string
                  nullable: true
                  example: "+250788000000"
      responses:
        '201':
          description: 'Employee created'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '422':
          description: 'Validation error'
  '/api/v1/employees/{id}':
    get:
      tags:
        - Employees
      summary: 'Show employee'
      operationId: b7cb6100cedcac60d8c23d3f9d7c3251
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Employee
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: 'Not found'
      security:
        - bearerAuth: []
    put:
      tags:
        - Employees
      summary: 'Update employee'
      operationId: bcdffcad24dd066f489c0e6ed83861a5
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Jane Doe
                email:
                  type: string
                  format: email
                  example: employee@example.com
                employee_identifier:
                  type: string
                  example: EMP-001
                phone_number:
                  type: string
                  nullable: true
                  example: "+250788000000"
      responses:
        '200':
          description: 'Employee updated'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: 'Not found'
        '422':
          description: 'Validation error'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Employees
      summary: 'Delete employee'
      operationId: 43aa0966dc246d0b0373a7db65a6bf78
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Employee deleted'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: 'Not found'
      security:
        - bearerAuth: []
  /api/v1/reports/attendance/daily:
    get:
      tags:
        - Reports
      summary: 'Daily attendance report (PDF or Excel)'
      operationId: 6fd430d445fbdd6d7a7f21ac45ebc183
      parameters:
        -
          name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        -
          name: format
          in: query
          required: false
          schema:
            type: string
            enum:
              - pdf
              - xlsx
      responses:
        '200':
          description: 'Report file stream'
        '401':
          description: Unauthenticated
      security:
        - bearerAuth: []
tags:
  -
    name: Attendance
  -
    name: Auth
  -
    name: Employees
  -
    name: Reports
    description: Reports
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "Token"